Resolve ember-data:default-serializers deprecation
