+++ addon/serializers/application.js
@@ -1,0 +1,1 @@
+export { default } from '@ember-data/serializer/json-api';
