+++ addon/serializers/version.js
@@ -1,0 +1,4 @@
+import JSONAPISerializer from '@ember-data/serializer/json-api';
+
+export default class VersionSerializer extends JSONAPISerializer {
+}
