+++ app/serializers/application.js
@@ -1,0 +1,1 @@
+export { default } from 'guidemaker/serializers/application';
