+++ app/serializers/version.js
@@ -1,0 +1,1 @@
+export { default } from 'guidemaker/serializers/version';
